feat: 공모 작성 API 구현 (#139)
* feat: 공모 작성 API 구현


* refactor: create를 save로 변경


* refactor: dto entity 매핑로직을 dto로 이동


* refactor: controller request 매개변수 명 컨벤션 적용


---------

+++ backend/src/main/java/com/zzang/chongdae/offering/domain/OfferingPrice.java
@@ -8,9 +8,11 @@ import lombok.AllArgsConstructor;
 public class OfferingPrice {
 
     private final Integer totalCount;
-    private final BigDecimal totalPrice;
+    private final Integer totalPrice;
 
-    public BigDecimal calculateDividedPrice() {
-        return totalPrice.divide(BigDecimal.valueOf(totalCount), RoundingMode.HALF_UP);
+    public Integer calculateDividedPrice() {
+        return BigDecimal.valueOf(totalPrice)
+                .divide(BigDecimal.valueOf(totalCount), RoundingMode.HALF_UP)
+                .intValue();
     }
 }
